refactor(localproxy): move unit tests co-located to src following __tests__ convention
