+++ packages/localproxy/src/services/caddyfile/__tests__/caddyfile.test.ts
@@ -3,7 +3,7 @@ import { fileURLToPath } from "node:url";
 import { describe, expect, it } from "vitest";
 import { CaddyfileService, type LocalDomain } from "#/services/caddyfile";
 import { type Caddyfile, CaddyfileParser } from "#/services/caddyfile/parser";
-import { FixtureReader } from "#test/helpers";
+import { FixtureReader } from "./helpers";
 
 const __dirname = path.dirname(fileURLToPath(import.meta.url));
 const fixtureReader = new FixtureReader(__dirname);
