+++ packages/localproxy/src/services/caddyfile/__tests__/fixtures/Caddyfile.simple
