+++ packages/localproxy/tsconfig.json
@@ -4,7 +4,6 @@
     "baseUrl": ".",
     "paths": {
       "#/*": ["./src/*"],
-      "#test/*": ["./test/*"],
       "@vlandoss/*": ["../../packages/*/src"]
     }
   }
